Remove reliance on map from Dijkstra algorithm
This allows Dijkstra maps to be created with any user defined map grid.
It also uses floating points now so that the user can have any type of
(2d) map they would like.

Also cleans up some of the code such as removing the "start" in the
RL_Path struct.